Get Beta version Ready
