+++ Assets/Imports/FPS/Scripts/Game/Managers/ObjectiveKillEnemies.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: acdcb0f3d139a45a0873b8cfffb41e3d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
