--- Assets/Imports/FPS/Scripts/Gameplay/Objectives/ObjectiveKillEnemies.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a3cff9918d8018a44989bf444e3232f3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
