+++ Assets/Prefabs/Monsters/VanillaMonster.prefab
@@ -5434,8 +5434,8 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: e27364eefd6dd0341ba72a3bfc0d9727, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  MaxHealth: 20
-  CriticalHealthRatio: 0.3
+  MaxHealth: 1
+  CriticalHealthRatio: 0
 --- !u!114 &6987078181750600552
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -5603,7 +5603,16 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   m_AgentType: 0
-  m_ScoopNeeded: 3
+  m_ScoopNeeded: 1
+  PawnSpeed: 10
+  AverageSpeed: 8
+  TankSpeed: 5
+  PawnHealth: 1
+  AverageHealth: 2
+  TankHealth: 3
+  PawnDamage: 1
+  AverageDamage: 2
+  TankDamage: 3
 --- !u!1 &6987078181903905783
 GameObject:
   m_ObjectHideFlags: 0
@@ -5890,7 +5899,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
 --- !u!95 &1946972988382532125
 Animator:
-  serializedVersion: 3
+  serializedVersion: 5
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
@@ -5903,7 +5912,9 @@ Animator:
   m_UpdateMode: 0
   m_ApplyRootMotion: 0
   m_LinearVelocityBlending: 0
+  m_StabilizeFeet: 0
   m_WarningMessage: 
   m_HasTransformHierarchy: 1
   m_AllowConstantClipSamplingOptimization: 1
-  m_KeepAnimatorControllerStateOnDisable: 0
+  m_KeepAnimatorStateOnDisable: 0
+  m_WriteDefaultValuesOnDisable: 0
