+++ Assets/Scripts/Agents/MonsterType.cs
@@ -23,6 +23,18 @@ public class MonsterType : MonoBehaviour
     [SerializeField] private EnemyType m_AgentType;
     [SerializeField] private FlavorType m_ScoopNeeded;
 
+    [SerializeField] private float PawnSpeed = 10.0f;
+    [SerializeField] private float AverageSpeed = 8.0f;
+    [SerializeField] private float TankSpeed = 5.0f;
+
+    [SerializeField] private float PawnHealth = 1.0f;
+    [SerializeField] private float AverageHealth = 2.0f;
+    [SerializeField] private float TankHealth = 3.0f;
+
+    [SerializeField] private int PawnDamage = 1;
+    [SerializeField] private int AverageDamage = 2;
+    [SerializeField] private int TankDamage = 3;
+
     [Tooltip("Max Health, Speed and Damage for this monster")]
     public MonsterStats Stats;
 
@@ -44,13 +56,13 @@ public class MonsterType : MonoBehaviour
         switch (m_AgentType)
         {
             case EnemyType.PAWN:
-                stats = new MonsterStats(1.0f, 10.0f, 1);
+                stats = new MonsterStats(PawnHealth, PawnSpeed, PawnDamage);
                 break;
             case EnemyType.AVERAGE:
-                stats = new MonsterStats(2.0f, 8.0f, 2);
+                stats = new MonsterStats(AverageHealth, AverageSpeed, AverageDamage);
                 break;
             case EnemyType.TANK:
-                stats = new MonsterStats(3.0f, 5.0f, 3);
+                stats = new MonsterStats(TankHealth, TankSpeed, TankDamage);
                 break;
         }
