+++ Assets/Scripts/UI/LoadSceneManager.cs
@@ -21,12 +21,20 @@ public class LoadSceneManager : MonoBehaviour
 
     public void MoveToScene()
     {
-        //SceneManager.LoadScene(SceneName);
+        SceneManager.LoadScene(SceneName);
     }
 
     //Added for a quick way to transition scenes on win
     public void MoveToScene(string SceneName)
     {
         //SceneManager.LoadScene(SceneName);
+    }
+
+    public void QuitGame()
+    {
+        #if UNITY_EDITOR
+            UnityEditor.EditorApplication.isPlaying = false;
+        #endif
+            Application.Quit();
     }
 }
