+++ Packages/manifest.json
@@ -1,14 +1,14 @@
 {
   "dependencies": {
-    "com.unity.collab-proxy": "1.15.13",
+    "com.unity.collab-proxy": "2.0.4",
     "com.unity.connect.share": "4.1.1",
-    "com.unity.ide.rider": "2.0.7",
-    "com.unity.ide.visualstudio": "2.0.14",
+    "com.unity.ide.rider": "3.0.21",
+    "com.unity.ide.visualstudio": "2.0.18",
     "com.unity.ide.vscode": "1.2.5",
     "com.unity.learn.iet-framework": "1.2.3",
     "com.unity.probuilder": "4.5.2",
     "com.unity.progrids": "3.0.3-preview.6",
-    "com.unity.test-framework": "1.1.31",
+    "com.unity.test-framework": "1.1.33",
     "com.unity.textmeshpro": "3.0.6",
     "com.unity.timeline": "1.4.8",
     "com.unity.ugui": "1.0.0",
