+++ ProjectSettings/EditorBuildSettings.asset
@@ -20,9 +20,9 @@ EditorBuildSettings:
   - enabled: 1
     path: Assets/Scenes/Final/TitleScreen.unity
     guid: 860bb3a665941427ca671757a9f78766
-  - enabled: 1
-    path: Assets/Scenes/Final/MainLevel.unity
-    guid: 1837155d4e24db54c98d8ff1e7e33bfb
+  - enabled: 0
+    path: 
+    guid: 00000000000000000000000000000000
   - enabled: 1
     path: Assets/Scenes/Final/WinScene.unity
     guid: f803a216a4dda4b01a82c968a8f6dd2f
@@ -32,4 +32,10 @@ EditorBuildSettings:
   - enabled: 0
     path: 
     guid: 00000000000000000000000000000000
+  - enabled: 1
+    path: Assets/Scenes/Final/EnemySandbox.unity
+    guid: 28a6d7609c47942ceb23aeebd0001708
+  - enabled: 1
+    path: Assets/Scenes/Final/MainLevel-LevelLayout.unity
+    guid: 1d078440c07780c43a12b5490c06c06c
   m_configObjects: {}
