+++ ProjectSettings/Packages/com.unity.connect.share/Settings.json
@@ -2,6 +2,12 @@
     "m_Name": "Settings",
     "m_Path": "ProjectSettings/Packages/com.unity.connect.share/Settings.json",
     "m_Dictionary": {
-        "m_DictionaryValues": []
+        "m_DictionaryValues": [
+            {
+                "type": "System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089",
+                "key": "buildOutputDirList",
+                "value": "{\"m_Value\":\"/Users/dotunoseni-adegbite/Unity/Builds/ScoopSlinger/beta-webgl/beta_814/beta_814;;;;;;;;;\"}"
+            }
+        ]
     }
 }
