+++ ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,2 @@
-m_EditorVersion: 2020.3.30f1
-m_EditorVersionWithRevision: 2020.3.30f1 (1fb1bf06830e)
+m_EditorVersion: 2020.3.48f1
+m_EditorVersionWithRevision: 2020.3.48f1 (b805b124c6b7)
